+++ sandpack-react/src/hooks/useErrorMessage.ts
@@ -1,5 +1,3 @@
-import * as React from "react";
-
 import { useSandpack } from "./useSandpack";
 
 /**
@@ -8,11 +6,6 @@ import { useSandpack } from "./useSandpack";
 export const useErrorMessage = (): string | null => {
   const { sandpack } = useSandpack();
   const { error } = sandpack;
-
-  React.useEffect(() => {
-    sandpack.errorScreenRegisteredRef.current = true;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   return error?.message ?? null;
 };